requirements.txt
# Core Django
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# Database
psycopg2-binary>=2.9.9

# Image handling
Pillow>=10.0.0
cloudinary>=1.36.0
django-cloudinary-storage>=0.3.0

# CORS
django-cors-headers>=4.3.0

# Filtering
django-filter>=23.3

# Rate limiting
django-ratelimit>=4.1.0

# Background tasks
celery>=5.3.4
redis>=5.0.0
django-redis>=5.4.0

# Email
sendgrid>=6.10.0

# Environment variables
python-decouple>=3.8

# Production server
gunicorn>=21.2.0
whitenoise>=6.6.0

# Database URL parsing
dj-database-url>=2.1.0

# Timezone support
pytz>=2023.3

# API documentation (optional)
drf-yasg>=1.21.7

# Testing
pytest>=7.4.3
pytest-django>=4.6.0
factory-boy>=3.3.0

# Code quality
flake8>=6.1.0
black>=23.11.0